+++ CHANGELOG.md
@@ -45,6 +45,7 @@ Section Order:
 
 ### Changed
 
+- Use SDE fields in admin
 - Migration dependency updates to squashed migrations in Alliance Auth v5
 
 ## [6.0.0] - 2026-06-07